fix: review hints implemented
