+++ renku/service/.env-example
@@ -2,6 +2,7 @@ REDIS_HOST=redis
 REDIS_PORT=6379
 REDIS_DATABASE=0
 REDIS_PASSWORD=
+CACHE_HEALTH_CHECK_INTERVAL=60
 
 # Server
 CACHE_DIR=
@@ -14,6 +15,8 @@ SERVICE_LOG_LEVEL=WARNING
 RENKU_SVC_WORKER_QUEUES=
 RENKU_SVC_CLEANUP_TTL_FILES=
 RENKU_SVC_CLEANUP_TTL_PROJECTS=
+WORKER_DATASET_JOBS_TIMEOUT=
+WORKER_DATASET_JOBS_RESULT_TTL=
 RQ_WORKER_LOG_LEVEL=WARNING
 
 # Scheduler
